+++ src/MetricBundle/Resources/docs/commands.md
