+++ src/MetricBundle/Resources/docs/integration.md
